Rename the "modernize" category to "syntax sugar"
