+++ src/linter/syntax_sugar/enum_to_const.h
