+++ test/linter/utils.h
@@ -7,28 +7,26 @@
 #define LINT_AND_FIX_FOR_READABILITY(document)                                 \
   sourcemeta::alterschema::Bundle bundle;                                      \
   sourcemeta::alterschema::add(                                                \
-      bundle, sourcemeta::alterschema::LinterCategory::Modernize);             \
-  sourcemeta::alterschema::add(                                                \
       bundle, sourcemeta::alterschema::LinterCategory::AntiPattern);           \
   sourcemeta::alterschema::add(                                                \
       bundle, sourcemeta::alterschema::LinterCategory::Simplify);              \
   sourcemeta::alterschema::add(                                                \
       bundle, sourcemeta::alterschema::LinterCategory::Redundant);             \
+  sourcemeta::alterschema::add(                                                \
+      bundle, sourcemeta::alterschema::LinterCategory::SyntaxSugar);           \
   bundle.apply(document, sourcemeta::jsontoolkit::default_schema_walker,       \
                sourcemeta::jsontoolkit::official_resolver);
 
 #define LINT_AND_FIX_FOR_ANALYSIS(document)                                    \
   sourcemeta::alterschema::Bundle bundle;                                      \
   sourcemeta::alterschema::add(                                                \
-      bundle, sourcemeta::alterschema::LinterCategory::Modernize);             \
-  sourcemeta::alterschema::add(                                                \
       bundle, sourcemeta::alterschema::LinterCategory::AntiPattern);           \
   sourcemeta::alterschema::add(                                                \
       bundle, sourcemeta::alterschema::LinterCategory::Simplify);              \
   sourcemeta::alterschema::add(                                                \
-      bundle, sourcemeta::alterschema::LinterCategory::Desugar);               \
-  sourcemeta::alterschema::add(                                                \
       bundle, sourcemeta::alterschema::LinterCategory::Redundant);             \
+  sourcemeta::alterschema::add(                                                \
+      bundle, sourcemeta::alterschema::LinterCategory::Desugar);               \
   bundle.apply(document, sourcemeta::jsontoolkit::default_schema_walker,       \
                sourcemeta::jsontoolkit::official_resolver);
